update local function
Since only the one with the reporter job uses the camera functions, I added his control to the loops. So the loops won't run in those who don't have this job. This way it will be more optimized.

I removed the AddEventListener parts.

I combined it with #8.

By the way, I added PR #3

Before : 0.06 ms(not in job) After: 0.02 ms in job.

5/10:Update.
- Add local function